Cleanup warnings/minor issues

+++ examples/SLIP_InteractiveServer/HTML.h
@@ -90,7 +90,7 @@ static const char main_html_p2[] PROGMEM =
 * This allows the HTML code to be modified as desired, with no need to change any other code
 */
 void sendPage(EthernetClient& _client, const char* _pointer, size_t size ){
-  for(int i=0; i<size;i++){
+  for(uint16_t i=0; i<size;i++){
     char c = pgm_read_byte(_pointer++);
     _client.write(&c,1);
   }
